Ensure document version itm_idx is updated
If we disable spring's open-in-view there will be no Hibernate session
associated with the StoredDocument object that will automatically
persist any modifications we make to it.

We now need to explicitly save the StoredDocument when we add the new
version to its list of document versions, which will cause itm_idx to be
set for the new version.